requirements.txt
addict==2.4.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.13.1
fsspec==2024.2.0
gast==0.4.0
idna==3.4
jax==0.4.18
Jinja2==3.1.3
MarkupSafe==2.1.5
mmcv-full==1.5.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tomli==2.2.1
torch==1.11.0+cu113
torch-dct==0.1.6
torchvision==0.12.0+cu113
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==1.26.13
yapf==0.43.0


